PRF compute working version 1.
This architecture has 2 PRF memories connected through a compute unit. The two PRF needs to have the same scheme with this implementation.
It has 3 stages of execution: in the first stage PRF 1 and 2 are initialized with the input matrix data.
In the second stage elements are read from PRF 1 doing parallel reads, are computed in parallel, and written to PRF 2 using parallel writes.
The compute kernel is statically defined at synthesis time, however the element to compute upon can be defined at runtime using a BRAM mapped to the HOST.
Each element of the mapped BRAM is a 64 bit data, the first 32 bits are ignored ( for now, due to a bug in maxcompiler ) the last 32 bits represent the following:
| 14 bits for i_coordinate | 14 bits for j_coordinate | 4 bits for access type (rectangle, row etc. )